Move snowflake queries
